Add the statistics layer, and correct §8.6's threshold procedure
Implements §8: kappa, alpha, calibration diagnostics, stratified bootstrap,
selective and labour curves, IPW weighting, split sealing, and
class-conditional threshold selection. 100% line and branch coverage on
stats/, mypy strict clean.

Two specification changes, both because measurement disagreed with the design
(§18: reality wins, and the design is updated in the same commit).

1. §8.6's fixed-sequence walk is replaced by a full grid scan with a
   Bonferroni correction over the grid. Fixed-sequence assumes the most
   conservative band is easiest to certify. It is the hardest: a high t_hi
   accepts only a handful of gold positives, and a handful of clean items is
   not evidence of controlled risk (zero errors out of six gives p = 0.74).
   Worse, narrowing the band keeps confidently-wrong positives while dropping
   moderately-confident correct ones, so positive-class risk can rise as
   coverage falls. On a simulated coder leaking 1-2% of positives, the walk
   certified 0% of trials where the corrected scan certified ~47% at full
   coverage. The default grid shrinks 45 -> 15 points because the correction
   now makes resolution costly.

2. §14.6's miscalibration control was wrong as written. The guarantee is
   distribution-free, so a monotone overconfident coder preserves item
   ranking and risk control survives unharmed; it is not rejected as a rule.
   What is true, and is now what the test asserts, is that overconfidence is
   never rewarded: at balanced prevalence it loses the interior band the
   honest coder needs and is rejected outright.

The negative control holds at every prevalence: an all-negative coder is
refused, which is the whole reason the guarantee is class-conditional.

Python floor moves 3.10 -> 3.11; current numpy and scipy both require >=3.12,
so a 3.10 floor would mean pinning to old numpy against §13.2's intent.

+++ src/scruple/stats/agreement.py
@@ -64,7 +64,7 @@ def cohens_kappa_from_confusion(cm: Confusion, *, n_items: int | None = None) ->
     order to judge how much evidence there is.
     """
     n = cm.n
-    reported_n = n_items if n_items is not None else int(round(n))
+    reported_n = n_items if n_items is not None else round(n)
     if n <= 0:
         return Estimate(value=None, n=reported_n, reason=ReasonCode.EMPTY_INPUT)
 
@@ -106,7 +106,7 @@ def krippendorff_alpha_from_confusion(cm: Confusion, *, n_items: int | None = No
         alpha = 1 - D_o / D_e = 1 - (b + c)(2N - 1) / (n_1 * n_0)
     """
     n = cm.n
-    reported_n = n_items if n_items is not None else int(round(n))
+    reported_n = n_items if n_items is not None else round(n)
     n_total = 2.0 * n
     if n_total <= 1.0:
         # Fewer than one pairable value: there is nothing to build a chance
